pyproject.toml
[tool.poetry]
name = "modu-property"
version = "0.1.0"
description = ""
authors = ["copyNdpaste <samnaka@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.18"
amqp = "5.1.1"
asgiref = "3.6.0"
async-timeout = "4.0.2"
attrs = "22.2.0"
backports-zoneinfo = "0.2.1"
billiard = "3.6.4.0"
black = "22.12.0"
celery = "5.2.7"
certifi = "2022.12.7"
charset-normalizer = "2.1.1"
click = "8.1.3"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.2.0"
django = "4.0.8"
django-celery-beat = "2.4.0"
django-cors-headers = "3.13.0"
django-environ = "0.9.0"
django-timezone-field = "5.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
exceptiongroup = "1.1.0"
execnet = "1.9.0"
flower = "1.2.0"
gunicorn = "20.1.0"
humanize = "4.4.0"
idna = "3.4"
iniconfig = "1.1.1"
kombu = "5.2.4"
mypy-extensions = "0.4.3"
packaging = "22.0"
pathspec = "0.10.3"
platformdirs = "2.6.0"
pluggy = "1.0.0"
prometheus-client = "0.15.0"
prompt-toolkit = "3.0.36"
psycopg2-binary = "2.9.5"
pyjwt = "2.6.0"
pytest = "7.2.0"
pytest-django = "4.5.2"
pytest-xdist = "3.1.0"
python-crontab = "2.7.1"
python-dateutil = "2.8.2"
pytz = "2022.7"
redis = "4.4.0"
requests = "2.28.1"
six = "1.16.0"
soupsieve = "2.3.2.post1"
sqlparse = "0.4.3"
tomli = "2.0.1"
tornado = "6.2"
typing-extensions = "4.4.0"
tzdata = "2022.7"
urllib3 = "1.26.13"
vine = "5.0.0"
wcwidth = "0.2.5"
pytest-dotenv = "^0.5.2"
xmltodict = "^0.13.0"
pymysql = "^1.1.0"
cryptography = "^41.0.4"
pre-commit = "^3.4.0"
manticoresearch = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
